requirements.txt
# Core FastAPI
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
gunicorn>=21.2.0

# Database
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
alembic>=1.13.0
supabase>=2.1.0

# Authentication & Security
pyjwt>=2.8.0
python-jose[cryptography]>=3.3.0
slowapi>=0.1.9
passlib[bcrypt]>=1.7.4

# Secret Management (Production)
boto3>=1.34.0  # AWS Secrets Manager
azure-keyvault-secrets>=4.7.0  # Azure Key Vault
azure-identity>=1.15.0  # Azure authentication
google-cloud-secret-manager>=2.16.0  # GCP Secret Manager

# Monitoring & Logging
prometheus-fastapi-instrumentator>=6.1.0
sentry-sdk>=1.38.0
python-json-logger>=2.0.7

# Caching
redis>=5.0.1

# HTTP & Async
httpx>=0.25.2
aiohttp>=3.9.1
requests>=2.31.0

# AI & ML
openai>=1.3.0
numpy>=1.25.0

# Testing
pytest>=8.4.2
pytest-cov>=4.1.0
pytest-asyncio>=1.2.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
psutil>=5.9.6
watchdog>=3.0.0

# Universal Design System
# Additional dependencies for enhanced functionality
python-multipart>=0.0.6
jinja2>=3.1.2
